added license mgr, spaces reader platform roles; moved platform roles mgmt to platform module; existing spaces have flag set; added flag to allow subspace admins to invite to parent space
