+++ src/common/enums/authorization.credential.ts
@@ -3,9 +3,11 @@ import { registerEnumType } from '@nestjs/graphql';
 // Credentials to be added later:
 export enum AuthorizationCredential {
   GLOBAL_ADMIN = 'global-admin', // able to do everything, god mode
-  GLOBAL_COMMUNITY_READ = 'global-community-read', // able to view all details of the top level community
   GLOBAL_SUPPORT = 'global-support', // able to manage platform level information, can per space have admin rights
+  GLOBAL_LICENSE_MANAGER = 'global-license-manager', // able to manage platform level information, can per space have admin rights
   GLOBAL_REGISTERED = 'global-registered', // credential issued to all registered users
+  GLOBAL_COMMUNITY_READ = 'global-community-read', // able to view all details of the top level community
+  GLOBAL_SPACES_READER = 'global-spaces-read', // able to view all details of the top level community
 
   USER_SELF_MANAGEMENT = 'user-self', // able to update a user
